pkgs/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
black==24.10.0
numpy==2.1.3
pyannote.audio==3.3.2
torch==2.5.1
torchaudio==2.5.1
soundfile==0.12.1
ffmpeg-python==0.2.0
huggingface_hub==0.25.2
